txncorrectnesstest: move dirty write test to own file
